requirements.txt
Pillow
pytesseract
pyzbar
pandas
###xlsxwriter
isbnlib
